src/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.25
werkzeug==2.0.1
requests==2.26.0
